Separate main function in main.go and mark as !test
This will make more easier to control the unit tests, since main is
not possible to test.
