+++ .github/workflows/test.yml
@@ -47,7 +47,7 @@ jobs:
     - name: Calc coverage
       run: |
         export PATH=$PATH:$(go env GOPATH)/bin
-        go test -v -covermode=count -coverprofile=coverage.out
+        go test -v -covermode=count -tags test -coverprofile=coverage.out
     - name: Convert coverage to lcov
       uses: jandelgado/gcov2lcov-action@v1.0.0
       with:
